Option added to generate the ctags.
When the script is used in the git hooks, it calls git itself, making a
concurrency issue between the two git processes checking for the
index.lock.

So the script checks if we are running in a git repo only if the
`check-git` option is provided, which we want in the case of running it
in Vim.