Adjust vim autocompletion.
The change is for the situation when the last word of the line is
autocompletable, but the user wants a new line.  Before, pressing ENTER
would select the menu item, making it hard to add a new line when the
menu is displayed, and an item is selected.  Now, if the menu is
displayed (because the first typed letters are candidate to
autocompletion), and the first item is selected, but not inserted in the
document, then pressing TAB completes it and pressing ENTER closes the
menu and adds a new line.